Unsetting MAX keyword and 15000 char limit for question (#133)
Reporting fix LC-709

+++ src/main/java/uk/gov/cshr/civilservant/dto/QuizDataTableDto.java
@@ -10,5 +10,5 @@ import lombok.*;
 public class QuizDataTableDto {
   private String profession;
   private int numberOfAttempts;
-  private float averageScore;
+  private Double averageScore;
 }
